+++ examples/redpanda_serde.py
@@ -1,9 +1,6 @@
 """
-Schema registry complete example.
-
-The Kafka input connector has support for schema registries.
-We support Redpanda and Confluent registries.
-This example shows how to use the clients in the kafka connector.
+Serialization and deserialization using redpanda's schema registry
+and plain avro format.
 
 The schems used for this example are the following:
 
@@ -50,7 +47,7 @@ import bytewax.operators as op
 import bytewax.operators.window as wop
 from bytewax.connectors.kafka import KafkaSinkMessage, KafkaSourceMessage
 from bytewax.connectors.kafka import operators as kop
-from bytewax.connectors.kafka.registry import ConfluentSchemaRegistry
+from bytewax.connectors.kafka.serde import PlainAvroDeserializer, PlainAvroSerializer
 from bytewax.dataflow import Dataflow
 from bytewax.operators.window import SystemClockConfig, TumblingWindow
 from confluent_kafka.schema_registry import SchemaRegistryClient
@@ -58,39 +55,32 @@ from confluent_kafka.schema_registry import SchemaRegistryClient
 logger = logging.getLogger(__name__)
 logging.basicConfig(format=logging.BASIC_FORMAT, level=logging.WARNING)
 
-
-# Setup some more env vars first:
 KAFKA_BROKERS = os.environ.get("KAFKA_SERVER", "localhost:19092").split(";")
-IN_TOPICS = os.environ.get("KAFKA_IN_TOPIC", "in_topic").split(";")
+IN_TOPICS = os.environ.get("KAFKA_IN_TOPIC", "in-topic").split(";")
 OUT_TOPIC = os.environ.get("KAFKA_OUT_TOPIC", "out_topic")
-CONFLUENT_URL = os.environ["CONFLUENT_URL"]
-CONFLUENT_USERINFO = os.environ["CONFLUENT_USERINFO"]
-CONFLUENT_USERNAME = os.environ["CONFLUENT_USERNAME"]
-CONFLUENT_PASSWORD = os.environ["CONFLUENT_PASSWORD"]
+REDPANDA_REGISTRY_URL = os.environ["REDPANDA_REGISTRY_URL"]
 
-# Pass this to both kafka_in and kafka_out
-add_config = {
-    "security.protocol": "SASL_SSL",
-    "sasl.mechanism": "PLAIN",
-    "sasl.username": CONFLUENT_USERNAME,
-    "sasl.password": CONFLUENT_PASSWORD,
-}
 
 flow = Dataflow("schema_registry")
-kinp = kop.input(
-    "kafka-in", flow, brokers=KAFKA_BROKERS, topics=IN_TOPICS, add_config=add_config
-)
+kinp = kop.input("kafka-in", flow, brokers=KAFKA_BROKERS, topics=IN_TOPICS)
+
 # Inspect errors and crash
 op.inspect("inspect-kafka-errors", kinp.errs).then(op.raises, "kafka-error")
 
-# ConfluentSchemaRegistry config:
-sr_conf = {"url": CONFLUENT_URL, "basic.auth.user.info": CONFLUENT_USERINFO}
-registry = ConfluentSchemaRegistry(SchemaRegistryClient(sr_conf))
+# Redpanda's schema registry configuration
+client = SchemaRegistryClient({"url": REDPANDA_REGISTRY_URL})
 
-# Confluent's deserializer doesn't need a schema, it automatically fetches it.
-key_de = registry.deserializer()
-val_de = registry.deserializer()
+# Use plain avro instead of confluent's wire format.
+# We need to specify the schema in the deserializer too here.
+key_schema = client.get_latest_version("sensor-key").schema
+key_de = PlainAvroDeserializer(schema=key_schema)
+
+val_schema = client.get_latest_version("sensor-value").schema
+val_de = PlainAvroDeserializer(schema=val_schema)
+
+# Deserialize both key and value
 msgs = kop.deserialize("de", kinp.oks, key_deserializer=key_de, val_deserializer=val_de)
+
 # Inspect errors and crash
 op.inspect("inspect-deser", msgs.errs).then(op.raises, "deser-error")
 
@@ -124,26 +114,22 @@ def calc_avg(key__wm__batch) -> KafkaSinkMessage[Dict, Dict]:
         value={
             "identifier": key,
             "avg": sum(batch) / len(batch),
-            "window_open": wm.open_time.isoformat(),
-            "window_start": wm.close_time.isoformat(),
+            "window_start": wm.open_time.isoformat(),
+            "window_end": wm.close_time.isoformat(),
         },
     )
 
 
 avgs = op.map("avg", windows, calc_avg)
+
 op.inspect("inspect-out-data", avgs)
 
-# The serializers require a specific schema instead. You can use SchemaRef too.
-key_ser = registry.serializer(100002)
-val_ser = registry.serializer(100003)
+key_ser = PlainAvroSerializer(schema=key_schema)
+out_val_schema = client.get_latest_version("aggregated-value").schema
+val_ser = PlainAvroSerializer(schema=out_val_schema)
+
 # Serialize
 serialized = kop.serialize("ser", avgs, key_serializer=key_ser, val_serializer=val_ser)
 
 op.inspect("inspect-serialized", serialized)
-kop.output(
-    "kafka-out",
-    serialized,
-    brokers=KAFKA_BROKERS,
-    topic=OUT_TOPIC,
-    add_config=add_config,
-)
+kop.output("kafka-out", serialized, brokers=KAFKA_BROKERS, topic=OUT_TOPIC)